fix gradle warnings
this one:
As part of making the publishing plugins stable, the 'deferred
configurable' behavior of the 'publishing {}' block has been deprecated.
In Gradle 5.0 the 'enableFeaturePreview('STABLE_PUBLISHING')' flag will
be removed and the new behavior will become the default. Please add
'enableFeaturePreview('STABLE_PUBLISHING')' to your settings file and do
 test run by publishing to a local repository. If all artifacts are
 published as expected, there is nothing else to do. If the published
 artifacts change unexpectedly, please see the migration guide for more
 details:
 https://docs.gradle.org/4.10.2/userguide/publishing_maven.html#publishing_maven:deferred_configuration.